requirements.txt
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
drf-spectacular==0.27.0
pydantic==2.5.0 
dependency-injector==4.48.2
requests==2.32.5
celery==5.5.3
